refactor, fuzz: Split out addrman utils
